Rename parsl.dataflow.error to errors, plural, for consistency
errors.py is more common in the existing codebase than error.py:

$ find parsl -name error\*.py | sed 's/^.*\(error.*\.py\)$/\1/' | sort | uniq  -c
      3 error.py
      6 errors.py
