+++ parsl/config.py
@@ -7,7 +7,7 @@ from typing_extensions import Literal
 from parsl.utils import RepresentationMixin
 from parsl.executors.base import ParslExecutor
 from parsl.executors.threads import ThreadPoolExecutor
-from parsl.dataflow.error import ConfigurationError
+from parsl.dataflow.errors import ConfigurationError
 from parsl.dataflow.taskrecord import TaskRecord
 from parsl.monitoring import MonitoringHub
