+++ parsl/tests/test_python_apps/test_fail.py
@@ -97,7 +97,7 @@ def test_deps(numtasks=2):
 
     try:
         print("Final status : ", fu_final.result())
-    except parsl.dataflow.error.DependencyError as e:
+    except parsl.dataflow.errors.DependencyError as e:
         print("Caught the right exception")
         print("Exception : ", e)
     except Exception as e:
